Bugfixes
Fixed some combat resets for all the npcs and added a reset event in order to avoid future problems when others try to start the quest. When Adyen, Karja and Orelis die the event is supposed to reset. Blizzard had alot of troubles with this quest and often GM's had to come and reset it. We should prepare it for every scenario. Right now the quest works if you aggro kaylaan and don't let him kill your 3 escorts. Only minor problems left.