Core/Movement: Fix warrior charge animation
**Description:**

Warrior's Charge animation (**for players around, but not for the warrior itself**) get very weird sometimes. The problem is when the facing of the caster changes together with the launch of the charge movement, OR even more often, while jumping backwards and then immediately launching charge. This happens with intervene as well, and probably more spells like these.

**Current behaviour:**

The warrior appears like teleporting, the charge animation starts, then warrior moves about 2-3 yards forward, then gets back, and immediately teleported to the target after that.

**Expected behaviour:**

Charge animation should be smooth.

**Steps to reproduce the problem:**

1. Make Warrior
2. Make another character as target
3. Jump backwards and immediately press charge

Charge animation always looks smooth on the warrior's client, but very weird and broken on the target's client (this is not always happening, im using a macro to reproduce it easily:

> /cast charge
> /target player
> .cool
> /targetlasttarget