Refine cache usage in gh-actions
Common keys block the efficient usage of cache in forks as when one build succeeds, other builds of the workflow or a later commit will match that cache with the common key. The stored cache may be for completely different compiler altogether, so cache is never hit.
Without the common key, caches for wrong compilers are not matched and caches are searched from the main repository, which contains recent caches for all of the compilers.