Changed to use functions and always reset to 3306 port before eval
We need to always reset to 3306 to catch cases, where the database in between is not set to use a specific port. This is due to the fact that the MYSQL_PORT is set and the eval with the new data will not overwrite it thus keeping the old port number instead of defaulting to 3306.