fix(CORE/MetaGems): Meta gems sometimes ignore added sockets.
Fix #3598

For optimization's sake, the code handling meta gems only consider items with sockets.
However, the check is incomplete and only look at built-in sockets and not those added by upgrades.
Therefore, added sockets are considered by meta gems only in the event that the item already had a built-in socket.