fix(DB/gameobject): The Damaged Chest quest Tome of the Cabal should respawn right away
The Damaged Chest has a 5 minute respawn rate - which makes questing as a warlock
painful if there are other warlocks on the same quest.

The best video I could find: https://youtu.be/5TXyVDWkaec?t=197 from Classic.

There are other videos but the seem to be from private servers with a 5 min spawn rate.

Others with a similar issue: https://github.com/FeenixServerProject/Archangel_2.4.3_Bugtracker/issues/3310