fix(CORE/CREATURE) Polymorph erratic mob behaviour
On the first polymorph cast on a creature, they are still trying to target the player.
This lasts the entire first cast of polymorph on a creature.
The sheep will stop moving and face the player after every spline path completes looking very strange.

There is many ways to fix this but I went with the cleanest solution I feel like.