fix(CORE/UNIT) Fix criticals not working against sitting players.
The issue is that we set the enemy player into the stand state when we put them into combat.
As a result, when we check to see if they are in the sit state later, its always false.
So the only way you were getting a critical was with the base crit chance.

With this fix if an enemy player is sitting. And you swing at them you will always crit.
This also removes the chance to miss or dodge when sitting down, which I believe is blizzlike, and saves me from passing around a ton of parameters,
or having to cache the state inside of the class.

https://github.com/azerothcore/azerothcore-wotlk/issues/5635