Update rev_1625489162833644602.sql
Just saw the info in the wiki about using a number > 50000 for gossip IDs that are not sniffed. I was using 11860, which is also the creature_templaye ID for this creature as it was free, so changing this to the first free number > 50K. Re-tested with new ID, still working.