fix(Core/Loot):
Players that did not participate in killing dungeon boss are not eligible to get loot.
Players that are too far away from the looted object are not eligible to get loot.
Players that released spirit and were outside the dungeon when the loot has been released are eligible to get loot.
Players that have pending bind are not eligible to get loot.
Properly get loot recipient for some chests in dungeons.
All aboved fixes should work in any loot mode (group loot, master loot, etc.)
Fixed #2104.