Fix incorrect pathing of caster mobs when their casts fail due to LOS
The original problem was that the casters, when their casts were blocked
by a los of line of sight, would not run directly towards the player, but
would rather run around within the spell range until a clear line of
sight was found. This made it near impossible to make casters stop their
casts and force them into melee range by using an obstacle.

This fix makes caster mobs run directly into melee until a clear line of
sight is reached whenever a player hides behind an obstacle regardless of
whether the mob is within spell range or not.