fix: Queue solo players, new damage formulas, and stop pet PvP crash
Fix for solo LFG given the base mechanics of the dungeon finder are that if the last person in the dungeon finder leaves the instance then the instance that was under the dungeon finder is no longer held nor are you in LFG anymore. If you are not in a party and you leave that instance (can be left by dying) the instance is no longer there. When you go to get your corpse it then treats it as you entering the instance from the portal location instead of through LFG as well. This is the exact behavior it would have if there was no LFG module and you were in a party of 5, 4 people left, then 1 person died and released spirit.

tl;dr, If you are solo there is no party.

Additionally we've worked out some new damage formulas that scale towards max players and a multiplier of one. This is a nerf to groups in dungeons but they needed a nerf given how much XP you can get. The new formulas make a lot more sense and should have been implemented in the first place.

Finally pets who got the killing blow on a player were crashing the server. This has now been fixed.