First attempt
so far it doesnt work, but here is what does work.
It doesnt throw a error up if there is a scriptname in the source but not assigned if you have the script name in the ScriptName of the creature table instead of the creature_template table. Progress i guess but not the end result we want. Npc info does not show the scripted ai being linked to the creature via guid.