Core/Objects: increase sight range of objects & correct general default visibility distance
This is a refactor and alteration of TC Cherry pick Commit  of https://github.com/TrinityCore/TrinityCore/pull/22891
This commit is designed to correct issue https://github.com/azerothcore/azerothcore-wotlk/issues/9175 as the gob will be loaded on the client before the npc. It will also do put us in line with tc with visibility distance column as well.