fix(DB/Quest): Noth the Plaguebringer ends quest 'Noth's special brew'
Noth the Plaguebringer should not reward the quest 12717 "Noth's special brew",
the player must complete the quest to the cauldron instead. Actually, both are
active: the player can end this quest by talking to Noth or to the cauldron.

Ported from TheLegendaryKingdom/LegacyOfAzeroth (not committed in own rep, directly edited on DB)

Ayahuasca (TLK)