Loot Correction
While Triaging https://github.com/chromiecraft/chromiecraft/issues/1893
I performed a creature loot table audit and discovered one item that is no present in any sniff or datamining.
The item in question is the Stranglekelp. This loot is not mentioned or showed in any sniff or tables dealing with the rare creature Lady Zephris https://tbc.wowhead.com/npc=14277/lady-zephris