fix(Core\OpCode): HandleFarSightOpcode (#9965)
* Change TO ThreadUNSAFE

This will process the far sight packet as a is not thread-safe - process it in World::UpdateSessions() compared to the former packet is thread-safe - process it in Map::Update().
This matches up with tc in everything we have with the HandleFarSightOpcode.

* Further Adjustments

* Move to Debug Loggin

TC has it as move to Debug Logging. Stating its a weird opcode handling issue which is not really known why.