fix(Core/Logs): improve log money (#10297)
* Fixed wrong decription of error

The Logger was pointing to the wrong table. Text is missing not in `broadcast_text_locale`, but in `broadcast_text`.

* fix(DB/Logger): fix logs topics and types

Correction of texts in the topic and the type of logged action. Getting rid of `<>` tags, which can cause certain difficulties when processing logs by the administrator. Now we can sort entries by the type of logged actions

* fix log_money

* fix log_money

* fix log_money

* fix log_money

* Update ObjectMgr.cpp

* Update rev_1642821969869259834.sql

* reduce memory usage

* Update AuctionHouseMgr.cpp

* Update Guild.cpp

* Update MailHandler.cpp

* Update TradeHandler.cpp

* fix typo

* Update data/sql/updates/pending_db_characters/rev_1642821969869259834.sql