fix (core): EffectInstaKill Prevents spell animation before deal damage execution
This fixes the issue of spell animations of chain effects of a spell dealing with EffectInstaKill. This was identified as a issue in support of @malcrom 's current scripting efforts. We fix the issue by putting a half second delay and making the delay of a function possible and independent.

Co-Authored-By: Malcrom <malcromdev@gmail.com>