Add (DB/SAI): Raider Jhash and Raider Kerr
This is a cherry pick of sorts from https://github.com/TrinityCore/TrinityCore/issues/27680 Which is authored by @alastro-Kraken and signed off by @offl . i am huge supporter of adding life role plays in the game, so it doesn't seem like a walking simulator, sometimes its about the trip and not the destination. However because our sai system is seriously lacking, jacked, busted on parts, and at times have issues with wanting to work with other tables which I have zero idea how or why, this had to be adjusted dramatically as sorts as the default sai from the TC did not work on AZ. So changes made on this end for AZ, we did not include creature formation table additions as for some reason creature formations and smartai scripted npcs do not cooperate with each other, which lead me to use action 29 to put them at a angle and distance, however for some reason our angle is opposite from tc so I have to adjust it to be 90 instead of 270 and change it from a distance of 3 to 2.5.  I had to add in a orientation adjustment for AZ as the JHash thing didn't keep its orientation for some reason on a waypoint pause at razor hill. The damn set data didn't work at all for raider kerr on azerothcore for him to move to the water point in razor hill so i had to make that a freaking timed action list, so I tossed in some 400ms timers so the damn thing will fire off. #400msClub as having it as zero for the time actions didn't fire off at all. I also had to set Raider Kerr to follow Raider Jhash on "Respawn" as that was the only way for it to fire off so he would even follow Jhash in the begin at Valley of Trials and mid point at razor hill, why am i forced to use on respawn for Kerr to follow Jhash after the razor hill stop, i have no idea, but it was the only way for the sai to fire off for kerr to follow jhash for some reason. The actions are then repeated as a loop so it will be seamless which that part was not adjusted. NPC Flag, extra flag, and gossip had to be adjusted for raider JHash as he was never a gossip npc and no exp is gained from him on kill.

Co-Authored-By: Alastro <98045288+Alastro-Kraken@users.noreply.github.com>
Co-Authored-By: offl <11556157+offl@users.noreply.github.com>