Modifying door closing event
From a logical point of view, when using the gossip, the door should close. Actually, before this change, the door closes, but only when the encounter is about to start. What I did, was to change the position, the place where the closing event is triggered. When the player, that the gossip, the door automatically closes. Preventing, that the event is activated and that the player can escape through the door.