Modifying door opening events
Just as we have modified the closing events to prevent the player from escaping at the beginning of the event, we must also control that the door opens in relation to the end of the different dialogues.

Initially, the variable gossipMenuID was initialized in the constructor, however, when using a character of the horde, in the last event, it did not show me any dialogue to begin the last battle. So I moved the initialization.