Avoiding events if another event is in progress
The idea is to prevent the gossip from being available until the previous event has ended. Therefore, I removed the lines where the npc's gossip was enabled ahead of time, and reordered it within the door opening event. In other words, the door and the npc are available at the same time. That way, we get the event to end, then the door opens and also enables the npc for the next dialog.