fix(CORE/Naxxramas): Portals and Teleports (#12696)
This fixes overlapping portals inside and outside Naxxramas. Their
rotation is also corrected (from CCW to CW). The post-boss
Naxxramas portals now teleport to the blue hub portal and activate only
when all wings are cleared. This requirement can be bypassed with the
orbs, which are now correctly
casting their respective sapphiron entry/exit spells in addition to having the
correct position.