Fixes
Removed the remaining hardcoded texts.
Added missing texts for the head.
Uncommented CloseGossipMenuFor since the gossip window doesn't close without it.
Removed the daily quest from Pumpkin Shrine. I can't find any evidence showing that it should exist but all videos I find from that era are only from the Cataclysm pre-patch and the quest isn't there in any of them. I believe the quest was was removed when the event was changed to only be accessible from the dungeon finder. If it's wrong, I'll remove that query.