Add (core\logging): Log sync db queries in mapupdater
This will Log sync db queries in World::Update() loop for the mapupdater when building in debug.
Originally part of TC Commit https://github.com/TrinityCore/TrinityCore/pull/25174
Apparently azerothcore never cherrypicked it entirely for whatever reason and is needed when running a debugger in debug build for me to find and address some recent crash issues that been posted on the issues to fully verify the issue.

Co-Authored-By: Giacomo Pozzoni <giacomopoz@gmail.com>