fix(Core/Pets) Pet scaling only being applied to pets with DB Entry
Original script was only applying the config option
`CONFIG_ALLOWS_RANK_MOD_FOR_PET_HEALTH`
to pets who did not have a DB entry. It was part of the else statement.

This PR creates a new if statement inside where the scaling value will be applied appropriately if this bool is true.