fix(DB/questpoi): 7321 -Soothing Turtle Bisque - correct POI for Horde
This fixes POI area for 7321 which shares it with the 555 (aliance version) which has the correct ones. Please beware of my SQL skill, although it fixes the problem, I encourage you to review this code and if necessarry, make it more compact. The basic idea is just to duplicate entries from 555 to 7321 - this results in correct behavior of the quest poi on the world map.