fix(Crash/Quest/Spell) : The Boon of Remulos
Fix the spell : https://www.wowhead.com/wotlk/spell=57678/emerald-acorn-effect
Used in quest : https://www.wowhead.com/wotlk/quest=13075/the-boon-of-remulos

The crashed was caused by a nullptr Excception. It also could be throuwn in another similar test that has also been fixed.
If the spell did not have a CastTimeEntry a nullptr Exception was not checked.