fix(DB/SAI): Missing scripting from Dark Rider and DeathCharger
Dark Rider of Acherus (id: 28768) was missing some on death scripting
that controled the deathcharger (id: 28782). Use smart ai to make
the deathcharger rooted in place, emote a message,
remove not selectable unit flag to allow player to mount
and apply 100% health to ensure deathcharger isn't at
50% health.

Added the emote text to creature_text table for deathcharger.

added NON_ATTACKABLE and NOT_SELECTABLE unitflags to deathcharger.

Added SmartAI Script to deathcharger, so that when a passenger enters
the horse unroots.

Closes AzerothCore issue #20107