fix(Core/SAI): rooted check blocks vehicles from casting quest credit spell
The quest SteamTank Surprise has a SmartAI entry in the db that causes
the tank to cast the kill credit spell on the player when they
successfully destroy a plague wagon. Currently, this spell cast is being
blocked by the rooted check handler for SMART_ACTION_CAST. The change
proposes that this rooted check should not apply to vehicles.

Closes AzerothCore issue #20237