fix(CORE/Spells): Add spell range correction for 43458 and 43468
Both the quests The Echo of Ymiron and Anguish of Nifflevar have spells
that are meant to be cast for credit at the end of the dialogues.
Because their range is set to 0 in Spell.dbc, this does not happen
unless you stand on top of the npcs that give credit. This fix adds
appropriate ranges to each spell.

Closes AzerothCore issue #19776
Closes AzerothCore issue #18874