Fix Gossip Waypoint Pause and Resume 12-15-24
This fixes the problem of NPCs and such that stop after a player opens a gossip window or quest but they do not resume their waypoint.  It uses both the CONFIG_WAYPOINT_MOVEMENT_STOP_TIME_FOR_PLAYER option in the worldserver conf and the interactive pause in the creature template movement table.  Also fixes the Resumewaypoint smartscript action where the delay didn't work.