fix(Scripts/AuchenaiCrypts): Remove debuff on summon death
Exarch Maladaar debuffs players when casting Stolen Soul, and summons a
Stolen Soul mob. When the mob dies, the debuff is supposed to be removed
from the player.

To achieve this, keep track of the number of Stolen Souls targeting each
player. If that number falls to 0, remove the debuff.

Closes AzerothCore issue #22419